+++ common_lib/Cargo.toml
@@ -6,16 +6,14 @@ edition = "2021"
 [dependencies]
 serde = "1.0.179"
 
-
-
 #serde = "1.0.179"
 serde_json = "1.0.104"
 #tracing-subscriber = { version = "0.3.17", features = ["json"] }
 
-#testcontainers = "0.15.0"
+testcontainers = "0.15.0"
 async-trait = "0.1.74"
 aws-sdk-lambda = "0.34.0"
-#aws-sdk-dynamodb = "0.34.0"
+aws-sdk-dynamodb = "0.34.0"
 #aws-config = "0.56.1"
 tracing = { version = "0.1.40", features = ["log"] }
 aws-sdk-sqs = "0.34.0"
@@ -29,3 +27,5 @@ aws-config = "0.56.1"
 # TODO check if we can get rid of it
 tokio = { version = "1", features = ["full"] }
 rand = "0.8.5"
+regex = "1.10.2"
+chrono = "0.4.31"
