+++ common_lib/manifest.test.snapstart.json
