+++ common_lib/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod reponse;
+pub mod report_log;
 pub mod retry_manager;
 pub mod runtime;
 
+pub mod dynamodb_manager;
 pub mod lambda_manager;
 pub mod manifest;
