+++ function-deployer-rs/src/cloudwatch_manager.rs
@@ -72,7 +72,7 @@ impl LogManager for CloudWatchManager {
             .put_subscription_filter()
             .destination_arn(&self.report_log_processor_arn)
             .filter_name(&format!("report-log-from-{}", runtime.function_name()))
-            .filter_pattern("REPORT")
+            .filter_pattern("REPORT RequestId")
             .log_group_name(format!("/aws/lambda/{}", runtime.function_name()))
             .send()
             .await?;
